fix(security): address gosec findings (G115/G703/G302/G108/G114/G404)
gosec now reports 0 issues (was 20). Changes:

G115 integer overflow casts (10 findings):
- pkg/draw/ximage.go: add X16/X32 clamp helpers for X11 uint16/uint32 ranges
- pkg/wmx11/fullscreen.go: use draw.X32 for window geometry
- pkg/xshm/xshm.go: use draw.X32 (shmid) and draw.X16 (w,h)
- pkg/apps/apps.go: make parseHex's hex() return uint8 (bounds 0-15)
- pkg/apps/uispec/uispec.go: clamp parsed color ints via clampByte()

G703/G302 path traversal + perms (6 findings):
- pkg/cmds/kitty_install.go: filepath.Clean the config dir; #nosec on the
  world-readable config writes (kitten/snippet are not secret, by design)

G108/G114 pprof auto-exposure + no http timeout (2 findings):
- pkg/cmds/wm.go: register pprof on a dedicated mux (not DefaultServeMux),
  use http.Server with ReadHeaderTimeout (WriteTimeout 0 for profile/trace)

G404 weak RNG (1 finding):
- pkg/apps/demoapps/colors.go: #nosec G404 (deterministic demo seed)

+++ pkg/apps/apps.go
@@ -130,21 +130,21 @@ func parseHex(s string) color.RGBA {
 	if len(s) != 7 || s[0] != '#' {
 		return draw.Faint
 	}
-	hex := func(c byte) int {
+	hex := func(c byte) uint8 {
 		switch {
 		case c >= '0' && c <= '9':
-			return int(c - '0')
+			return c - '0'
 		case c >= 'a' && c <= 'f':
-			return int(c-'a') + 10
+			return c - 'a' + 10
 		case c >= 'A' && c <= 'F':
-			return int(c-'A') + 10
+			return c - 'A' + 10
 		}
 		return 0
 	}
 	return color.RGBA{
-		R: uint8(hex(s[1])<<4 | hex(s[2])),
-		G: uint8(hex(s[3])<<4 | hex(s[4])),
-		B: uint8(hex(s[5])<<4 | hex(s[6])),
+		R: hex(s[1])<<4 | hex(s[2]),
+		G: hex(s[3])<<4 | hex(s[4]),
+		B: hex(s[5])<<4 | hex(s[6]),
 		A: 0xff,
 	}
 }
